Remove the minimum set of plugins required to get the app to work
The wikitude plugin doesn't really need to be removed, but we clearly don't
need it, so I removed it. And downgrading the push plugin to the one the
emission currently uses removed the other error, which had to do with nanop,
which is a small implementation of the google ProtoBuf standard.